+++ test/cases/msg-exception-execute-template/webpack.config.js
@@ -15,6 +15,4 @@ module.exports = {
   },
 
   plugins: [new PugPlugin({})],
-
-
 };