+++ test/cases/require-scripts-html/expected/assets/js/76.94ceb6bd.js
@@ -1,0 +1,1 @@
+"use strict";(self.webpackChunk=self.webpackChunk||[]).push([[76],{76:(e,s,n)=>{var c=n(817);const k=n.n(c)().render("SCRIPT B");console.log(k)}}]);