+++ test/cases/split-chunk-resolve-assets-prod/expected/assets/js/552.a8c7475e.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[171,552],{171:(o,l,e)=>{const n=e(198),s=e(740);console.log(">> main-a:"),console.log(" - A: ",n),console.log(" - C: ",s)}}]);