+++ test/cases/split-chunk-resolve-assets-prod/expected/assets/js/552.ee7f5099.js
@@ -1,1 +1,0 @@
-(self.webpackChunk=self.webpackChunk||[]).push([[552,171],{171:(o,l,e)=>{const n=e(198),s=e(740);console.log(">> main-a:"),console.log(" - A: ",n),console.log(" - C: ",s)}}]);