+++ test/cases/split-chunk-vendor/expected/assets/js/fixture-lorem.9010aee6.js
@@ -1,0 +1,1 @@
+"use strict";(self.webpackChunk=self.webpackChunk||[]).push([[928],{},s=>{s.O(0,[96],(()=>s(s.s=696))),s.O()}]);