+++ test/cases/split-chunk-vendor/expected/assets/js/fixture-lorem.f64d404c.js
@@ -1,1 +1,0 @@
-"use strict";(self.webpackChunk=self.webpackChunk||[]).push([[928],{},s=>{s.O(0,[96],(()=>(696,s(s.s=696)))),s.O()}]);